Fix AspectRatio not being calculated due to integer division

+++ Assets/Scripts/RealCamera/AttachCameraFeed.cs
@@ -30,7 +30,7 @@ public class AttachCameraFeed : MonoBehaviour
 
         if (AutoAlign)
         {
-            var aspectRatio = new Vector2(imageWidth / imageHeight, -1);
+            var aspectRatio = new Vector2((float)(imageWidth) / (float)(imageHeight), -1);
             transform.localScale = new Vector3(aspectRatio.x, aspectRatio.y, 1.0f);
 
             float xOffset = 0;
